requirements.txt
aiogram==2.11.2
aiohttp>=3.6.2
aioredis==1.3.1
python-dotenv~=0.15.0

gino~=1.0.1
SQLAlchemy~=1.3.22
asyncpg~=0.21.0